Handle more player loading logic on the server
Have the server handle figuring out starting transforms so the spawn
points get distributed properly. Also adds a thing where the player
position average gets offset by an average of the players' original
starting points, which (mostly) fixes the problem where the level
wouldn't start flat. A wait for all clients to be ready and synced still
needs to be added before rotating the level for this to work smoothly.